Change read_csv C++ API into a class method.
