+++ cpp/CMakeLists.txt
@@ -257,7 +257,7 @@ add_library(cudf SHARED
             src/io/utilities/parsing_utils.cu
             src/io/utilities/file_utils.cpp
             src/io/convert/dlpack/cudf_dlpack.cpp
-            src/io/csv/csv_reader.cu
+            src/io/csv/csv_reader_impl.cu
             src/io/csv/csv_writer.cu
             src/io/json/json_reader_impl.cu
             src/io/orc/orc_reader.cu
